Moody/fee options array (#70)
* suggested change

* another simplification

* use loops everywhere

* add todo

+++ contracts/libraries/PriceMath.sol
@@ -87,7 +87,7 @@ library PriceMath {
     function getInputToRatio(
         uint112 reserveIn,
         uint112 reserveOut,
-        uint16 lpFee,
+        uint24 lpFee,
         FixedPoint.uq112x112 memory inOutRatio
     )
         internal
